Complete - task : Debug Alloy API
Incomplete - task : Matching of outcomes

Increased memory limit for command line tests, but allow scenarios with TiedSoreLosers to be skipped if solution is not found, rather than increase scope until memory is exhausted. This is because some scenarios involving TiedSoreLosers are not possible.  Added skeleton models for risk and trust based on previous work for the Fair Trade case study.